update api for locationTable and table

+++ restaurant/src/main/java/fu/rms/controller/TableController.java
@@ -3,12 +3,10 @@ package fu.rms.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import fu.rms.dto.TableDto;
-import fu.rms.mapper.TableMapper;
 import fu.rms.service.ITableService;
 
 @RestController
@@ -24,7 +22,7 @@ public class TableController {
 	}
 
 	@PutMapping("/table")
-	public TableDto updateStatusTable(@RequestParam Long tableId, int status) {
+	public TableDto updateStatusTable(@RequestParam Long tableId, Long status) {
 		return tableService.updateStatus(tableId, status);
 	}
